Changes rule_id to long in order to have it visible by systemtap

+++ src/rule.h
@@ -47,7 +47,7 @@ class Rule {
 
     std::vector<Variables::Variable *> *variables;
     int phase;
-    double rule_id;
+    long rule_id;
 
     Rule *chainedRule;
     bool chained;
